Refactor some

+++ db/migrate/20181001120814_create_reviews.rb
@@ -1,7 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[5.1]
   def change
     create_table :reviews do |t|
-      t.integer :score
+      t.float :score
       t.string :body
       t.integer :user_id
       t.integer :artist_id
